Drop RnRS PDFs from the archive
We discovered that files.scheme.org is not a good place for
unversioned files that keep changing. Every filename in the archive
should have a version number or a date (with the possible exception of
"-latest" files that are HTTP redirects to the latest version of a
file, e.g. "mosh-latest.tar.gz" redirecting to "mosh-0.2.7.tar.gz").

Henceforth, the RnRS PDFs are available from
https://standards.scheme.org/ (among other places around the net) and
a new "scheme-standards" project will release bundles of them, with
the latest bundle always extracted onto standards.scheme.org. The
bundles will be archived (with dated filenames) on files.scheme.org.

One file remains for now: "r2rs-nbs.pdf". It's not clear what to do
about the pre-R3RS standards for which we don't have TeX sources.

Originally discussed with @weinholt months ago, postponed until now.